Update from PR comments
